Add a way to create a variable length transaction in the DAML scenario.
The complex scenario will now send a variable length transaction to the ledger api service.

Sample usage:
python3 daml_requests.py --url https://1.2.3.4:6865 scenario --complex --exec-delay-factor 30

The delay factor triggers a calculation loop in the DAML code.  The ledger api
will time out at 30 seconds, which is typically met/exceeded with an exec-delay-factor of 33-34.
The default is 1.

Not related to the new code: I am commenting out the creation of an Asset.Clerk in the complex
scenario. The complex scenario was always hanging.  When I kill with Ctrl-C, this message
is given:

Writer loop for party Charlie has NOT fully finished, but will be terminated anyway (1 futures still pending).

Charlie is the Clerk.  Commenting out this line allows the scenario to finish.