TRC: Refactor TRS connection
The thin replica client object handles an array of already initialized gRPC
stubs. The stubs themselves are initialized with a gRPC channel which is the
object that maintains the actual connection, however, the TRC has no access to
this channel. Meaning, we have no direct access to the underlying connection
and cannot explicitly connect or disconnect ourselves. With this change, we
abstract the gRPC connection into a separate object which is the only way for
the TRC to interact with a TRS. Note, the connection object API only deals with
protobuf messages and hides the gRPC backend completely.
This will allow us, in future changes, to connect and disconnect from a TRS and
simplify testing.