BC-4941: Make user accept EULA at login
[1] When root or blockchain user logs in, display the EULA text, and
prompt for yes|no. On no, exit.

On yes, create a file under /opt/vmware/blockchain/eula with the
name .$USER.eula. Record the time of acceptance in this file.

On subsequent logins check for the file. If it exists, proceed
normally with login. If it does not exist, show the eula text, etc.

[2] Switch the the firstboot mechanism provided by the CAP folks.
We were using a home-grown firstboot service to perform firstboot
tasks. This is more easily achieved by the CAP-provided firstboot
functionality.