Add default configuration for concord5 in docker
The config distribution script
(docker/config-public/distribute-configuration-files.sh) expects all
concord directories to be created. If any of them is missing - the
distribution fails. This patch adds default RO config for concord5.

If for some setups concord5 is not RO replica the config will be
overwritten with the correct one so there will be no issues.