distribute-configuration-files.sh logs warning on missing dir
If any of the source/destination directories is missing on the mv step
of the script - log on stderr.
This is workaround for the cases when the script should distribute
configuration for replica count different from 4. The destination dir
for the script is hard-coded in hermes and can't be adjusted easily.