DAML: Remove color from logs
While colors are great, they make reading logs hard if the color code doesn't
get interpreted properly. In logback, you can define log formats for every
output and it is very common to use colors on stdout. However, we also store
the output from stdout into a file and later use it for analysis. Meaning, the
logs that we look at contain ANSI color codes which makes it harder to read.
This change removes the colors in order to improve log analysis. Note, proper
log reader tools (such as lnav) will introduce colors if it sees a known log
format.