TRS: Update sync'ing
Recently, we replaced the circular buffer with a bounded queue. Meaning, we
don't need care about the queue limit anymore. As soon as the producer hits the
limit it won't be able to add updates anymore and drops them instead. The
consumer will be notified as soon as he tries to read from the queue. In any
case, it doesn't change the behavior of how we sync between the kvb and the
live-update queue.