BC-4495: Fix on-prem provisioning-allocation certs failure
A recent change introduced a requirement for provisioning service  to
communicate with IPAM only over TLS. For on-prem deployments in
production (at customer sites), IPAM will not be used because customers
will provide explicit IP addresses for their blockchain VMs.

So provisioning <=> IPAM communication is not necessary. However, the
castor prereqs file needs to be updated to set the protocol to NONE,
otherwise provisioning service bringup fails.