BC-4637: Log read set hash in DAML pre-execution
In some cases we observe certain replicas calculate results that don't match. This could be explained by the fact that they start from different block heights. To confirm this (and to exclude DAML EE non-determinism), we provide a DEBUG log statement that prints a digest of the read set as a whole.