Add CIDs to Thin Replica Client Logs
This commit adds correlation IDs for received updates to the Thin
Replica Client's Logs (when such IDs are available). The IDs are added
to the log context when a data update is received and kept as long as
that data update is being processed and validated via an RAII-like
mechanism implemented with the thin_replica_client::LogCid class
(introduced in this commit and declared in thin_replica_client.hpp.