TRC: Correct debug message
We recently refactored the TRS connections. At the end of every loop we check
that we don't have too many hash streams open. However, without this change the
debug message will be printed whether we actually close the stream or not.
This change makes sure that the debug message makes sense again.