BC-5180: Update docker image base version in Castor IT
The Castor IT deploys blockchains and then validates their
functionality. It delegates to helper functions to run daml sanity
tests. However, for the tests to succeed, the docker images need to
be at the right version. The sanity test calls are similar to those in
PersephoneSmokeNew, but PersephoneSmokeNew also sets up the docker
base image version. The same thing also needs to be done for
CastorDeploymentsTest.