BC-5235: Print consortium name and id to output file
There is a request to print the consortium ID to the output file.
Currently, with Castor, users provide the consortium name. A random UUID
is generated and sent to provisioning service.

This id is important because it is used during logging, and this
information will be necessary to filter the logs during analysis.

Print out the consortium name and generated ID in the Castor output
file.