BC-4981: Remove obsolete OldJwtSecurityConfig
This removes the old JwtSecurityConfig.  As a result, we no longer need the flag vmbc.auth.csp.  It is always true.

There were several unit tests based on this code that have also been removed.

There is still some cleanup work to do, to get rid of other aspects of the old authorization system.

No need for JwtTokenFilter.java