Unit test correct RPC usage by the TRC
This commit adds the Thin Replica Client unit testing scoped by the
BC-1931 Jira task. In particular, this commit unit tests that the
ThinReplicaClient implementation uses the Thin Replica RPC calls as
intended by the Thin Replica mechanism design. Though Thin Replica
Servers need to be able to handle Byzantine-faulty clients, we do not
intend for our production implementation of the Thin Replica Client to
be Byzantine-faulty out-of-the-box, so we expect it to only use the RPC
calls as designed.

This commit also includes additions to the Thin Replica Server mocking
declared in thin_replica_client_mocks.hpp and implemented in the
corresponding cpp file. In particular, it adds the
ThinReplicaCommunicationRecord and MockThinReplicaServerRecorder
classes, which are used to record and scrutinize which RPC calls the
ThinReplicaClient makes to the mock servers in the unit tests.