BC-5363: Expose the telegraf metrics pull port to the outside world
The endpoint/port at which the telegraf prometheus_client output plugin
serves requests needs to be accessible by the outside world, so it
needs to listen on all interfaces. Otherwise, clients residing on other
machines will not be able to connect to this endpoint.

Add 0.0.0.0 as the host ip to which to bind the 9273 port.