Add TRS-TRC performance benchmark tool
This commit adds the performance benchmark tool. This includes
mocks of TRC and TRS. The TRC mock is built on top of the trc-client
library. The TRS mock is built on top of the concord image.
The benchmark parameters are specified in an environment file,
which is loaded by a docker-compose file. Given a number of blocks,
key value pairs, key size and value size, the tool generates dummy data
based on the values of the parameters specified.
These values are then read by the clients, when it subscribes to a server.