BC-5363: Enable metrics pull for on-prem deployments - Part III
This MR is part III of III.

Users may specify a username and password in the infrastructure
descriptor (and later, Helen) for the Telegraf Prometheus client output
plugin. If the credentials are specified, that information makes its
way via Castor -> Provisioning Service -> Config Service -> Agent,
where the agent then modifies the telegraf config file to enable the
output plugin and set up credentials for it.

Part I:
Configuration Service + agent side changes to allow the
username/password to be stored in CS, to be picked up by the agents.
Agent modifies the telegraf config file on the appliance to enable
the prometheus-client plugin with the specified username/password. This
launces an HTTP endpoint at port 9273 for incoming pull-based calls.
The endpoint is available under http://<IP>:9273/metrics.

Part II:
Castor + Provisioning side changes: Castor takes (optional)
username/password properties for the output plugin, and passes it to PS.

Part III:
Enable TLS for https between the consumer of the pull API. If the user
has provided the key and cert for the prometheus_client output plugin,
then enable those fields in the telegraf conf template. Also set up one
ConfigurationComponent for the key file and one for the cert file in
ConfigurationServiceHelper. This configuration is then used by the agent
to create those directories+files, and copy the contents specified into
them.