Fix replica id logging in concord log file
Effectively the replica ID for the committer node overlaps with its
index in node configuration array (node section from the config). This
feature was used to log the replica id via its index and its actual
value over the config file.

With the introduction of RO replica, this situation changes. RO replicas
are defined in separate section and they have got separate indexes,
which are different from their actual replica ids.

For example we have got 4 commiters and 1 RO node. The committers are
defined in nodes section and their indexes are 0, 1, 2 and 3. They are
exactly the same as their replica ids.
However RO node is defined in ro_node section with index 0, but it's
actual replica id is 4.

This patch fixes the logs which sometimes showed indexes and sometimes -
replica ids. When replica id is logged it is obtained from the
configuration. For the cases when index is logged - this is explicitly
stated.