Move certain code out of configuration_manager.cpp
This commit moves implementations of ConcordConfiguration::ParameterValidator,
ConcordConfiguration::ParameterGenerator, ConcordConfiguration::ScopeSizer,
and ParameterSelection::ParameterSelector out of configuration_manager.cpp
to a separate header file; however, this header file is not intended for
inclusion anywhere other than configuration_manager.cpp. This change is
made in the interest of reducing the length of configuration_manager.cpp
in order to improve code navigability.