Change the defaut application.config file for concord with static one
application.config contains default values for the concord replicas. It
is generated from a predefined file during concord-core image building.
This file generate dummy node sections from a template and adds private
parameters for them.

This is fine for committer nodes because the configuration merging
algorithm handles the private parameters. However the RO replicas are
not handled by the algorithm and the default file breaks the RO
deployments with Helen and Castor.

This patch changes the file from auto-generated to static one with just
the required global parameters. It is compatible with RO and committer
replicas.

This patch is meant to be an emergency fix before a proper mitigation is
implemented.