BC-5739: Fix invalid API authentication check call
[1] The validation code I added to check if a vCenter is alive and
authenicated used an API,

    /api/vcenter/namespaces/instances

which is not available on all VCs (only those that have namespaces
enabled).

So the API call fails causing the deployment to fail on such VCs. Fix
this by switching to a standard API,

    /rest/vcenter/certificate-management/vcenter/tls

that is available on all VCs regardless of configuration.

[2] The VALIDATE option skipped the run through javax validation, so it
missed catching schema validation errors. Fix this.