Fix config generation for RO replica with use_loopback_for_local_hosts
During config generation the committer and RO nodes are identified by
their indexes in node and ro_node sections. As these are two separate
sections there are cases when a RO replica idx is the same as a
committer node idx. E.g. 1st commiter has got idx=0 in nodes and 1st RO
replica has idx=0 in ro_nodes.

This causes some confusion when use_loopback_for_local_hosts is enabled
in configuration input file. This option changes the IP address of the
current replica to 127.0.0.1 so that not to use public address if it
needs to communicate with itself (keep in mind that when docker is used
a lot of traffic redirection occurs and the IP address in config is not
necessary the same as the IP address of the container).

Due to the overlapping indexes and the nature of opereation of
use_loopback_for_local_hosts, when a config for a RO replica is
generated the IP address of the committer node with idx equal to the RO
idx is set to 127.0.0.1. This is a problem because first it is wrong,
second the RO replica effectively doesn't know how to communicate with
this committer.

This is fixed by adding a check in outputConcordNodeConfiguration()
which indicates if a config for RO replica is generated.