Authorize TRC at TRS
This commit adds authorization of TRC when TLS is enabled for TRC-TRS.
When a connecting client's client id is to be fetched, instead of
fetching it from the metadata, TRS obtains it from the OU field
in the certificate provided by the TRC. TRS then compares the
client_id with the client_id it knows from its cumulative root
certificate. If the client_ids match the connecting client is authorized.