BC-5739: Fix invalid API authentication check call
Fix for a fix:

The validation code I added to check if a vCenter is alive and
authenicated used an API,

    /rest/vcenter/certificate-management/vcenter/tls

which is not available on all VCs at version 6.7.0.

So the API call fails causing the deployment to fail on such VCs. Fix
this by switching to a standard API,

    /rest/com/vmware/cis/tagging/category

that *seems* to be available on all VCs that I have tried so far.