BC-5889: Make container registry mandatory
There's been a change in the PS recently where the hardcoded container
registry information was removed, since it is a security risk.

This has caused deployments via Castor to fail if the infra descriptor
does not have the container registry specified. With this change in PS,
it now becomes mandatory to specify the registry properties in the zone
specification in the infra descriptor.

Validate that the containerRegistry property is always specified.