Fix BC-5879
This fixes BC-5879. For TRC-TRS unique certificate folder names
are only required for the docker-compose environment and a
generic path can be used by the deployed containers, since
each node needs only its own certs and the bundled root cert.
This commit updates the docker-compose files to mount the
local folder names to generic cert path input by the user for
both TRC and TRS. It also updates the TRS and TRC to consume
this generic cert path name.