BC-5960: Collect support bundles on test failure.
The current behavior is to only collect support bundles:
- If a test failed and the blockchain nodes to use were passed in on the cmd line.
- If deployment fails.

The new behavior is to also collect the support bundle if we did a deployment and a test fails.

Needed as part of BC-5960.

Also removed extraneous, unused support bundle code.  This code path appears to have been
inactive for a while.