Operator: Refactor request timeout
Checking for a valid timeout parameter is needed for all endpoints. This change
refactors timeout handling into a separate function. It covers an additional
error case and aligns the time unit. If the timeout is not valid then the
helper function will prepare the response as well so that this doesn't have to
be duplicated either. In addition, the concop script now provides a generic
timeout parameter instead of one for each subcommand. While touching the file
we also align the formatting.